+++ frontend/lib/src/data/models/cart_line.dart
@@ -4,13 +4,13 @@ class CartLine {
   const CartLine({required this.variant, required this.quantity});
 
   final ProductVariant variant;
-  final int quantity;
+  final double quantity;
 
   double get subtotal => variant.unitPrice * quantity;
 
   double get total => subtotal;
 
-  CartLine copyWith({int? quantity}) {
+  CartLine copyWith({double? quantity}) {
     return CartLine(variant: variant, quantity: quantity ?? this.quantity);
   }
 }
