+++ frontend/lib/src/features/invoices/view_models/invoice_details_view_model.dart
@@ -109,7 +109,7 @@ class InvoiceDetailsViewModel extends ChangeNotifier {
       order: order,
       reason: reason,
       metrics: {
-        'returned_quantity': lines.fold<int>(
+        'returned_quantity': lines.fold<double>(
           0,
           (sum, line) => sum + line.quantity,
         ),
