+++ frontend/lib/src/features/register_sessions/view_models/register_session_history_view_model.dart
@@ -339,7 +339,7 @@ class RegisterSessionHistoryViewModel extends ChangeNotifier {
       order: order,
       reason: reason,
       metrics: {
-        'returned_quantity': lines.fold<int>(
+        'returned_quantity': lines.fold<double>(
           0,
           (sum, line) => sum + line.quantity,
         ),
